Handle SIGTSTP and SIGCONT properly in client (#29)
This patch sets our a "dirty" flag on our client renderer so that calls
to the render from other running threads do not write via ncurses while
the terminal is in a bad state following SIGCONT. It does this through a
lightweight wrapper around the default ncurses SIGTSTP handler.

Known bugs: For some reason, this patch only seems to work the first
time CTRL-Z (SIGTSTP) is pressed. I think there is something in the
ncurses handler that is messing with our handler. This can be addressed
in a follow-up patch.